+++ gleam_stdlib/CHANGELOG.md
@@ -6,7 +6,10 @@
   `unique`, and `zip` functions.
 - `map_dict:Map` renamed to `map_dict:MapDict`.
 - `map_dict` module gains `drop`, and `take` functions.
-- `str` module gains `append` function.
+- `str` module gains `append` function and loses `from_int`, `parse_int`,
+  `from_float`, `parse_float`, and `base_from_int`.
+- `int` module created with `parse`, `to_string`, and `to_base_string`.
+- `float` module created with `parse`, and `to_string`.
 
 ## v0.1.1 - 2019-04-17